no longer crashes with bad args on mute/unmute

+++ commands/mute.js
@@ -10,45 +10,51 @@ exports.main = function(bot, msg, timeout, botPerm, userPerm) { // Export comman
 	if (timeout.check(msg.author.id, msg)) {
 		return;
 	}
-	// Check for cooldown, if on cooldown notify user of it and abort command execution.
-	let args = msg.content.substr(config.commandPrefix.length + command.length + 2);
-	let lengthMute = parseInt(args.trim().split(' ')[1])*1000;
-	let arg2 = lengthMute;
-	let users = msg.guild.members;
-	let mutee = users.get(msg.mentions.users.first().id);
-	//console.log(msg.mentions.users.first());
-	let muted = msg.guild.roles.find("name", 'Muted').id;
-	var epoch = (new Date).getTime();
-	//var epoch_unmute = lengthMute + epoch;
-	
-
-	//TODO figure out ^^
-	try {
-		if (!msg.member.hasPermission("KICK_MEMBERS")) {
-			msg.reply("U R NOT A MODERATOR");
-		} else {
-			if (mutee.roles.has(muted)) {
-				msg.reply("that user is already muted");
-			} else {
-				//mute them.
-				mutee.addRole(muted);
-				msg.guild.createChannel('temp', 'voice').then(channel => {
-					mutee.setVoiceChannel(channel).then(member => {
-						channel.delete();
-					});
-				});
-				lengthMute = lengthMute+Date.now();
-				// fs.writeFile('serverconf/muted/' + mutee.id, msg.guild.id+' '+lengthMute, (err) => {
-				// 	if (err) {
-				// 		msg.reply("error writing to file, err: " + err.toString());
-				// 	}
-				// });
-				database.addMuted(mutee.id, msg.guild.id, lengthMute);
-				msg.reply("muted " + mutee.toString() + ' for ' + arg2 / 1000 + ' seconds.');
-			}
+	if(typeof msg.mentions.users.first() !== 'undefined'){
+		// Check for cooldown, if on cooldown notify user of it and abort command execution.
+		let args = msg.content.substr(config.commandPrefix.length + command.length + 2);
+		let lengthMute = parseInt(args.trim().split(' ')[1])*1000;
+		if(typeof lengthMute == 'undefined'||isNaN(lengthMute)){
+			lengthMute = 0;
+			console.log('undef lm');
 		}
-	} catch (err) {
-		msg.reply("failed, " + err.toString());
+		let arg2 = lengthMute;
+		let users = msg.guild.members;
+		let mutee = users.get(msg.mentions.users.first().id);
+		//console.log(msg.mentions.users.first());
+		let muted = msg.guild.roles.find("name", 'Muted').id;
+		var epoch = (new Date).getTime();
+		//var epoch_unmute = lengthMute + epoch;
+		
+
+		//TODO figure out ^^
+		try {
+			if (!msg.member.hasPermission("KICK_MEMBERS")) {
+				msg.reply("U R NOT A MODERATOR");
+			} else {
+				if (mutee.roles.has(muted)) {
+					msg.reply("that user is already muted");
+				} else {
+					//mute them.
+					mutee.addRole(muted);
+					msg.guild.createChannel('temp', 'voice').then(channel => {
+						mutee.setVoiceChannel(channel).then(member => {
+							channel.delete();
+						});
+					});
+					lengthMute = lengthMute+Date.now();
+					// fs.writeFile('serverconf/muted/' + mutee.id, msg.guild.id+' '+lengthMute, (err) => {
+					// 	if (err) {
+					// 		msg.reply("error writing to file, err: " + err.toString());
+					// 	}
+					// });
+					database.addMuted(mutee.id, msg.guild.id, lengthMute);
+					msg.reply("muted " + mutee.toString() + ' for ' + arg2 / 1000 + ' seconds.');
+				}
+			}
+		} catch (err) {
+			msg.reply("failed, " + err.toString());
+		}
 	}
 };
